backend/requirements.txt
Django
psycopg2-binary
gunicorn
djangorestframework
markdown
django-filter
django-cors-headers
Celery
redis